Separated master and worker code
