+++ service-runner.js
@@ -10,20 +10,9 @@
 // Upgrade to es6
 require('core-js/shim');
 
-// Use bluebird internally. Use P.resolve(es6Promise) to convert an incoming
-// Promise to a bluebird Promise.
-var P = require('bluebird');
-
 var cluster = require('cluster');
-var path = require('path');
-var yaml = require('js-yaml');
-var fs = P.promisifyAll(require('fs'));
-var os = require('os');
-
-var Logger = require('./lib/logger');
-var makeStatsD = require('./lib/statsd');
-var HeapWatch = require('./lib/heapwatch');
-var docker = require('./lib/docker');
+var Master = require('./lib/master');
+var Worker = require('./lib/worker');
 
 // Disable cluster RR balancing; direct socket sharing has better throughput /
 // lower overhead. Also bump up somaxconn with this command:
@@ -31,537 +20,17 @@ var docker = require('./lib/docker');
 cluster.schedulingPolicy = cluster.SCHED_NONE;
 
 function ServiceRunner(options) {
-    this.options = this._getOptions(options);
-
-    this._config = null;
-    this._logger = null;
-    this._metrics = null;
-
-    // Figure out the base path
-    this._basePath = /\/node_modules\/service-runner$/.test(__dirname) ?
-        path.resolve(__dirname + '/../../') : path.resolve('./');
-
-    // Is the master shutting down?
-    this._shuttingDown = false;
-
-    // Are we performing a rolling restart?
-    this._inRollingRestart = false;
-
-    this.workerHeartbeatTime = {};
+    if (cluster.isMaster) {
+        this._impl = new Master(options);
+    } else {
+        this._impl = new Worker(options);
+    }
 }
 
 ServiceRunner.prototype.run = function run(conf) {
-    var self = this;
-    var configUpdateAction;
-    if (cluster.isMaster) {
-        configUpdateAction = self.updateConfig(conf);
-    } else {
-        configUpdateAction = new Promise(function(resolve, reject) {
-            var timeout = setTimeout(function() {
-                reject(new Error('Timeout waiting for config in worker ' + process.pid));
-            }, 3000);
-            process.on('message', function(message) {
-                if (message.type === 'config') {
-                    clearTimeout(timeout);
-                    self.updateConfig(message.body)
-                    .then(resolve);
-                } else {
-                    reject(new Error('Invalid message received: ' + JSON.stringify(message)));
-                }
-            });
-        });
-    }
-    return configUpdateAction
-    .then(function() {
-        var config = self.config;
-        // display the version
-        if (self.options.displayVersion) {
-            console.log(config.serviceName + ' ' + config.package.version);
-            process.exit(0);
-        }
-
-        // do we need to use Docker instead of starting normally ?
-        if (self.options.useDocker) {
-            self.options.basePath = self._basePath;
-            return docker(self.options, self.config);
-        }
-
-        self._logger = new Logger(config.logging);
-
-        if (cluster.isMaster && self.config.num_workers > 0) {
-            return self._runMaster();
-        } else {
-            return self._runWorker();
-        }
-    });
+    return this._impl.run(conf);
 };
 
-ServiceRunner.prototype._sanitizeConfig = function(conf, options) {
-    // TODO: Perform proper validation!
-    if (!conf.logging) { conf.logging = {}; }
-    if (!conf.metrics) { conf.metrics = {}; }
-    // check the number of workers to run
-    if (options.num_workers !== -1) {
-        // the number of workers has been supplied
-        // on the command line, so honour that
-        conf.num_workers = options.num_workers;
-    } else if (conf.num_workers === 'ncpu' || typeof conf.num_workers !== 'number') {
-        // use the number of CPUs
-        conf.num_workers = os.cpus().length;
-    }
-    conf.worker_heartbeat_timeout = conf.worker_heartbeat_timeout || 7500;
-    return conf;
-};
-
-
-/**
- * Loads the config from file, serialized input or Object
- *
- * @param conf a configuration.
- *             If undefined - config is loaded from config.yaml file
- *             If Object - treated as already parsed configuration
- *             If sting - treated a serialized yaml config
- */
-ServiceRunner.prototype.loadConfig = function loadConfig(conf) {
-    var self = this;
-    var action;
-    if (conf && conf instanceof Object) {
-        // Ready config object, no need to load from FS or parse yaml
-        action = P.resolve(conf);
-    } else if (conf && typeof conf === 'string') {
-        // Yaml source provided as config string
-        action = P.try(function() {
-            return yaml.safeLoad(conf);
-        });
-    } else {
-        // No config provided - load from file and parse yaml.
-        var configFile = this.options.configFile;
-        if (/^\./.test(configFile)) {
-            // resolve relative paths
-            configFile = path.resolve(self._basePath + '/' + configFile);
-        }
-        action = fs.readFileAsync(configFile)
-        .then(function(yamlSource) {
-            return yaml.safeLoad(yamlSource);
-        });
-    }
-
-    var package_json = {};
-    try {
-        package_json = require(self._basePath + '/' + 'package.json');
-    } catch (e) {}
-
-    return action.then(function(config) {
-        config = self._sanitizeConfig(config, self.options);
-        config.package = package_json;
-        if (config.info) {
-            // for backwards compat
-            var pack = config.package;
-            pack.name = config.info.name || pack.name;
-            pack.description = config.info.description || pack.description;
-            pack.version = config.version || pack.version;
-        }
-        self.config = config;
-    })
-    .catch(function(e) {
-        console.error('Error while reading config file: ' + e);
-        process.exit(1);
-    });
-};
-
-/**
- * Updates the config and sets instance properties.
- *
- * @param conf a configuration.
- *             If undefined - config is loaded from config.yaml file
- *             If Object - treated as already parsed configuration
- *             If sting - treated a serialized yaml config
- */
-ServiceRunner.prototype.updateConfig = function updateConfig(conf) {
-    var self = this;
-
-    return self.loadConfig(conf)
-    .then(function() {
-        var config = self.config;
-        var name = config.package && config.package.name || 'service-runner';
-        config.serviceName = name;
-
-        // Set up the logger
-        if (!config.logging.name) {
-            config.logging.name = name;
-        }
-
-        // And the statsd client
-        if (!config.metrics.name) {
-            config.metrics.name = name;
-        }
-    });
-};
-
-ServiceRunner.prototype._checkHeartbeat = function() {
-    var self = this;
-    self.interval = setInterval(function() {
-        if (!self._shuttingDown && !self._inRollingRestart) {
-            var now = new Date();
-            Object.keys(cluster.workers).forEach(function(workerId) {
-                var worker = cluster.workers[workerId];
-                var lastBeat = self.workerHeartbeatTime[worker.process.pid];
-                if (!lastBeat || (!lastBeat.killed && now - lastBeat.time
-                        > self.config.worker_heartbeat_timeout)) {
-                    self._logger.log('error/service-runner/master',
-                        'worker ' + worker.process.pid + ' stopped sending heartbeats, killing.');
-                    self._stopWorker(workerId);
-                    // Don't need to respawn a worker, it will be restarted upon 'exit' event
-                }
-            });
-        }
-    }, self.config.worker_heartbeat_timeout / 2 + 1);
-};
-
-ServiceRunner.prototype._runMaster = function() {
-    var self = this;
-    // Fork workers.
-    this._logger.log('info/service-runner', 'master(' + process.pid + ') initializing '
-            + this.config.num_workers + ' workers');
-
-    cluster.on('exit', function(worker, code, signal) {
-        if (!self._shuttingDown && !self._inRollingRestart) {
-            var exitCode = worker.process.exitCode;
-            self._logger.log('error/service-runner/master',
-                    'worker ' + worker.process.pid
-                    + ' died (' + exitCode + '), restarting.');
-            delete self.workerHeartbeatTime[worker.process.pid];
-            P.delay(Math.random() * 2000)
-            .then(function() {
-                self._startWorkers(1);
-            });
-        }
-    });
-
-    var shutdown_master = function() {
-        self._shuttingDown = true;
-        self._logger.log('info/service-runner/master',
-                'master shutting down, killing workers');
-        if (self.interval) {
-            clearInterval(self.interval);
-        }
-        cluster.disconnect(function() {
-            self._logger.log('info/service-runner/master', 'Exiting master');
-            process.exit(0);
-        });
-    };
-
-    process.on('SIGINT', shutdown_master);
-    process.on('SIGTERM', shutdown_master);
-
-    // Set up rolling restarts
-    process.on('SIGHUP', function() {
-        return self.updateConfig()
-        .then(function() {
-            // Recreate loggers
-            self._logger.close();
-            self._logger = new Logger(self.config.logging);
-        })
-        .then(self._rollingRestart.bind(self));
-    });
-
-    return this._startWorkers(this.config.num_workers)
-    .then(function(workers) {
-        self._checkHeartbeat();
-        return workers;
-    });
-};
-
-ServiceRunner.prototype._stopWorker = function(workerId) {
-    var self = this;
-    var worker = cluster.workers[workerId];
-    self.workerHeartbeatTime[worker.process.pid] = {
-        time: null,
-        killed: true
-    };
-    var res = new P(function(resolve) {
-        var timeout = setTimeout(function() {
-            // worker.kill doesn't send a signal immediately, it waits until
-            // worker closes all connections with master. If after a minute
-            // it didn't happen, don't expect it happen ever.
-            process.kill(worker.process.pid, 'SIGKILL');
-            delete self.workerHeartbeatTime[worker.process.pid];
-            resolve();
-        }, 60000);
-        worker.on('disconnect', function() {
-            worker.kill('SIGKILL');
-            clearTimeout(timeout);
-            delete self.workerHeartbeatTime[worker.process.pid];
-            resolve();
-        });
-    });
-    worker.disconnect();
-    return res;
-};
-
-ServiceRunner.prototype._rollingRestart = function() {
-    this._logger.log('info/service-runner/master', {
-        message: 'SIGHUP received, performing rolling restart of workers'
-    });
-    var self = this;
-    self._inRollingRestart = true;
-    P.each(Object.keys(cluster.workers), function(workerId) {
-        return self._stopWorker(workerId)
-        .then(function() {
-            return self._startWorkers(1);
-        });
-    })
-    .then(function() {
-        self._inRollingRestart = false;
-    });
-};
-
-/*
-    This is a workaround for the following bug in node:
-    https://github.com/joyent/node/issues/9409
-
-    Shortly, the problem is that a worker is removed by master
-    if it sends 'disconnect' and then it's checked on 'exit'.
-    However, in some cases (for example an infinite CPU loop in worker,
-    'disconnect' may never happen, and master will crash receiving an 'exit'
- */
-function fixCloseDisconnectListeners(worker) {
-    // Take the exit listener node's cluster have set.
-    var listeners = worker.process.listeners('exit')[0];
-    var exit = listeners[Object.keys(listeners)[0]];
-
-    // Take the disconnect litener node's cluster have set.
-    listeners = worker.process.listeners('disconnect')[0];
-    var disconnect = listeners[Object.keys(listeners)[0]];
-
-    // Now replace the exit listener to make sure 'disconnect' was called
-    worker.process.removeListener('exit', exit);
-    worker.process.once('exit', function(exitCode, signalCode) {
-        if (worker.state != 'disconnected') {
-            disconnect();
-        }
-        exit(exitCode, signalCode);
-    });
-}
-
-/**
- * Saves the timestamp of a worker heartbeat
- * @private
- */
-ServiceRunner.prototype._saveBeat = function(worker) {
-    var self = this;
-    var currentVal = self.workerHeartbeatTime[worker.process.pid];
-    if (currentVal && currentVal.killed) {
-        return;
-    }
-    self.workerHeartbeatTime[worker.process.pid] = {
-        time: new Date(),
-        killed: false
-    };
-};
-
-// Fork off one worker at a time, once the previous worker has finished
-// startup.
-ServiceRunner.prototype._startWorkers = function(remainingWorkers) {
-    var self = this;
-    if (remainingWorkers) {
-        var worker = cluster.fork();
-        self._saveBeat(worker);
-        return new P(function(resolve) {
-            fixCloseDisconnectListeners(worker);
-            worker.send({
-                type: 'config',
-                body: yaml.safeDump(self.config)
-            });
-            worker.on('message', function(msg) {
-                if (msg.type === 'startup_finished') {
-                    resolve(self._startWorkers(--remainingWorkers));
-                } else if (msg.type === 'heartbeat') {
-                    self._saveBeat(worker);
-                }
-            });
-        });
-    }
-};
-
-ServiceRunner.prototype._workerHeartBeat = function() {
-    // We send heart beat 3 times more frequently than check it
-    // to avoid possibility of wrong restarts
-    this.interval = setInterval(function() {
-        process.send({ type: 'heartbeat' });
-    }, this.config.worker_heartbeat_timeout / 3);
-};
-
-ServiceRunner.prototype._runWorker = function() {
-    var self = this;
-    // Worker.
-    process.on('SIGTERM', function() {
-        if (self.interval) {
-            clearInterval(self.interval);
-        }
-        self._logger.log('info/service-runner/worker', 'Worker '
-                + process.pid + ' shutting down');
-        process.exit(0);
-    });
-
-    // Enable heap dumps in /tmp on kill -USR2.
-    // See https://github.com/bnoordhuis/node-heapdump/
-    // For node 0.6/0.8: npm install heapdump@0.1.0
-    // For 0.10: npm install heapdump
-    process.on('SIGUSR2', function() {
-        var heapdump = require('heapdump');
-        var cwd = process.cwd();
-        console.error('SIGUSR2 received! Writing snapshot.');
-        process.chdir('/tmp');
-        heapdump.writeSnapshot();
-        process.chdir(cwd);
-    });
-
-    self._metrics = makeStatsD(self.config.metrics, self._logger);
-
-    // Heap limiting
-    // We try to restart workers before they get slow
-    // Default to something close to the default node 2g limit
-    var limitMB = parseInt(self.config.worker_heap_limit_mb) || 1500;
-    new HeapWatch({ limitMB: limitMB },
-            this._logger,
-            this._metrics).watch();
-
-    if (cluster.isWorker) {
-        self._workerHeartBeat();
-    }
-
-    // Require service modules and start them
-    return P.all(this.config.services.map(function(service) {
-        var modName = service.module || service.name;
-        if (/^\./.test(modName)) {
-            // resolve relative paths
-            modName = path.resolve(self._basePath + '/' + modName);
-        }
-        var svcMod;
-        try {
-            svcMod = require(modName);
-        } catch (e) {
-            e.moduleName = modName;
-            return P.reject(e);
-        }
-
-        var opts = {
-            config: service.conf,
-            logger: self._logger.child({
-                name: service.name || service.module,
-            }),
-            // todo: set up custom prefix
-            metrics: self._metrics
-        };
-
-        return P.try(function() {
-            return svcMod(opts);
-        });
-    }))
-    .then(function(res) {
-        // Signal that this worker finished startup
-        if (cluster.isWorker) {
-            process.send({type: 'startup_finished'});
-        }
-        return res;
-    })
-    .catch(function(e) {
-        self._logger.log('fatal/service-runner/worker', e);
-        process.exit(1);
-    });
-};
-
-ServiceRunner.prototype._getOptions = function(opts) {
-
-    if (opts) {
-        // no need to parse command-line args,
-        // opts are already here
-        return opts;
-    }
-
-    // check process arguments
-    var args = require('yargs')
-        .usage('Usage: $0 [command] [options]')
-        .command('docker-start', 'starts the service in a Docker container')
-        .command('docker-test', 'starts the test process in a Docker container')
-        .command('build', 'builds the service\'s package and deploy repo')
-        .options({
-            n: {
-                alias: 'num-workers',
-                default: -1,
-                describe: 'number of workers to start',
-                nargs: 1
-            },
-            c: {
-                alias: 'config',
-                default: './config.yaml',
-                describe: 'YAML-formatted configuration file',
-                type: 'string',
-                nargs: 1
-            },
-            f: {
-                alias: 'force',
-                default: false,
-                describe: 'force the operation to execute',
-                type: 'boolean'
-            },
-            d: {
-                alias: 'deploy-repo',
-                default: false,
-                describe: 'build only the deploy repo',
-                type: 'boolean'
-            },
-            r: {
-                alias: 'review',
-                default: false,
-                describe: 'send the patch to Gerrit after building the repo',
-                type: 'boolean'
-            },
-            verbose: {
-                default: false,
-                describe: 'be verbose',
-                type: 'boolean'
-            },
-            v: {
-                alias: 'version',
-                default: false,
-                describe: 'print the service\'s version and exit',
-                type: 'boolean'
-            },
-        })
-        .help('h')
-        .alias('h', 'help')
-        .argv;
-
-    args.deployRepo = args.deployRepo || args.review;
-    args.build = args._.includes('build') || args.deployRepo;
-    args.dockerStart = args._.includes('docker-start');
-    args.dockerTest = args._.includes('docker-test');
-    args.deployRepo = args.deployRepo || args.build;
-
-    if (args.build && args.dockerStart || args.build && args.dockerTest
-            || args.dockerStart && args.dockerTest) {
-        console.error('Only one command can be specified!');
-        process.exit(1);
-    }
-
-    opts = {
-        num_workers: args.numWorkers,
-        configFile: args.config,
-        displayVersion: args.v,
-        build: args.build,
-        buildDeploy: args.deployRepo,
-        sendReview: args.review,
-        dockerStart: args.dockerStart,
-        dockerTest: args.dockerTest,
-        useDocker: args.deployRepo || args.dockerStart || args.dockerTest,
-        force: args.force,
-        verbose: args.verbose
-    };
-
-    return opts;
-};
 
 module.exports = ServiceRunner;
